requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
mysql-connector-python
PyJWT
